updated jetty with latest JETTY_VERSION=9.4.30.v20200611

+++ build.sh
@@ -1,6 +1,6 @@
-#!/bin/bash -x
+#!/bin/bash
 
-set -x
+set -e
 
 # Reference: 
 # - https://docs.docker.com/engine/userguide/containers/dockerimages/
@@ -9,24 +9,58 @@ set -x
 if [ $# -lt 1 ]; then
     echo "-------------------------------------------------------------------------------------------"
     echo "Usage: "
-    echo "  ${0} [<Dockerfile> <imageTag>]"
+    echo "  ${0} [<Dockerfile> <imageTag> [<some more optional arguments...>] ] "
     echo "e.g."
     echo "  ./build.sh ./centos/Dockerfile.centos.xfce.vnc openkbs/centos-xfce-vnc --no-cache  --build-arg OS_TYPE=centos'"
     echo "  ./build.sh ./Dockerfile.ubuntu.xfce.vnc openkbs/ubuntu-xfce-vnc --no-cache  --build-arg OS_TYPE=centos'"
     echo "-------------------------------------------------------------------------------------------"
 fi
-MY_DIR=$(dirname "$(readlink -f "$0")")
 
 DOCKERFILE=${1:-./Dockerfile}
-DOCKERFILE=$(realpath $DOCKERFILE)
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+    # Linux
+    MY_DIR=$(dirname "$(readlink -f "$0")")
+    DOCKERFILE=$(realpath $DOCKERFILE)
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+    # Mac OSX
+    MY_DIR=`pwd`
+else
+    MY_DIR=`pwd`
+fi
+
+BUILD_CONTEXT=$(dirname ${DOCKERFILE})
 
 imageTag=${2}
 
-if [ $# -get 2 ]; then
+if [ $# -gt 2 ]; then
     shift 2
     options="$*"
 else 
     options=""
+fi
+
+##########################################################
+#### ---- Whether to remove previous build cache ---- ####
+#### ---- Valid value: 0 (No remove); 1 (yes, remove)
+##########################################################
+REMOVE_CACHE=0
+
+###############################################################################
+###############################################################################
+###############################################################################
+#### ---- DO NOT Change the code below UNLESS you really want to !!!!) --- ####
+#### ---- DO NOT Change the code below UNLESS you really want to !!!!) --- ####
+#### ---- DO NOT Change the code below UNLESS you really want to !!!!) --- ####
+###############################################################################
+###############################################################################
+###############################################################################
+
+##########################################################
+#### ---- Generate remove cache option if needed ---- ####
+##########################################################
+REMOVE_CACHE_OPTION=""
+if [ ${REMOVE_CACHE} -gt 0 ]; then
+    REMOVE_CACHE_OPTION="--no-cache --rm"
 fi
 
 ###################################################
@@ -35,24 +69,37 @@ fi
 ORGANIZATION=openkbs
 
 ###################################################
-#### ---- Detect docker ----
+#### ---- Detect Docker Run Env files ----
 ###################################################
-DOCKER_ENV_FILE="./.env"
-function detectDockerEnvFile() {
+
+function detectDockerBuildEnvFile() {
     curr_dir=`pwd`
-    if [ -s "./.env" ]; then
-        echo "--- INFO: ./.env Docker Environment file (.env) FOUND!"
-        DOCKER_ENV_FILE="./.env"
+    if [ -s "${DOCKER_ENV_FILE}" ]; then
+        echo "--- INFO: Docker Build Environment file '${DOCKER_ENV_FILE}' FOUND!"
     else
-        echo "--- INFO: ./.env Docker Environment file (.env) NOT found!"
-        if [ -s "./docker.env" ]; then
-            DOCKER_ENV_FILE="./docker.env"
+        echo "*** WARNING: Docker Build Environment file '${DOCKER_ENV_FILE}' NOT found!"
+        echo "*** WARNING: Searching for .env or docker.env as alternative!"
+        echo "*** --->"
+        if [ -s "./docker-build.env" ]; then
+            echo "--- INFO: ./docker-build.env FOUND to use as Docker Run Environment file!"
+            DOCKER_ENV_FILE="./docker-build.env"
         else
-            echo "*** WARNING: Docker Environment file (.env) or (docker.env) NOT found!"
+            if [ -s "./.env" ]; then
+                echo "--- INFO: ./.env FOUND to use as Docker Run Environment file!"
+                DOCKER_ENV_FILE="./.env"
+            else
+                echo "--- INFO: ./.env Docker Environment file (.env) NOT found!"
+                if [ -s "./docker.env" ]; then
+                    echo "--- INFO: ./docker.env FOUND to use as Docker Run Environment file!"
+                    DOCKER_ENV_FILE="./docker.env"
+                else
+                    echo "*** WARNING: Docker Environment file (.env) or (docker.env) NOT found!"
+                fi
+            fi
         fi
     fi
 }
-detectDockerEnvFile
+detectDockerBuildEnvFile
 
 ###################################################
 #### ---- Container package information ----
@@ -71,12 +118,14 @@ BUILD_ARGS="--build-arg BUILD_DATE=${BUILD_DATE} --build-arg VCS_REF=${VCS_REF}"
 
 ## -- ignore entries start with "#" symbol --
 function generateBuildArgs() {
-    for r in `cat ${DOCKER_ENV_FILE} | grep -v '^#'`; do
-        echo "entry=> $r"
-        key=`echo $r | tr -d ' ' | cut -d'=' -f1`
-        value=`echo $r | tr -d ' ' | cut -d'=' -f2`
-        BUILD_ARGS="${BUILD_ARGS} --build-arg $key=$value"
-    done
+    if [ "${DOCKER_ENV_FILE}" != "" ] && [ -s "${DOCKER_ENV_FILE}" ]; then
+        for r in `cat ${DOCKER_ENV_FILE} | grep -v '^#'`; do
+            echo "entry=> $r"
+            key=`echo $r | tr -d ' ' | cut -d'=' -f1`
+            value=`echo $r | tr -d ' ' | cut -d'=' -f2`
+            BUILD_ARGS="${BUILD_ARGS} --build-arg $key=$value"
+        done
+    fi
 }
 generateBuildArgs
 echo "BUILD_ARGS=${BUILD_ARGS}"
@@ -100,21 +149,37 @@ function generateProxyArgs() {
     if [ "${NO_PROXY}" != "" ]; then
         PROXY_PARAM="${PROXY_PARAM} --build-arg NO_PROXY=\"${NO_PROXY}\""
     fi
+    if [ "${http_proxy}" != "" ]; then
+        PROXY_PARAM="${PROXY_PARAM} --build-arg http_proxy=${http_proxy}"
+    fi
+    if [ "${https_proxy}" != "" ]; then
+        PROXY_PARAM="${PROXY_PARAM} --build-arg https_proxy=${https_proxy}"
+    fi
+    if [ "${no_proxy}" != "" ]; then
+        PROXY_PARAM="${PROXY_PARAM} --build-arg no_proxy=\"${no_proxy}\""
+    fi
     BUILD_ARGS="${BUILD_ARGS} ${PROXY_PARAM}"
 }
 generateProxyArgs
-echo "BUILD_ARGS=> ${BUILD_ARGS}"
+echo -e "BUILD_ARGS=> \n ${BUILD_ARGS}"
+echo
 
 ###################################################
-#### ---- Build Container ----
+#### ----------- Build Container ------------ #####
 ###################################################
 
+cd ${BUILD_CONTEXT}
 set -x
-echo "========> imageTag: ${imageTag}"
-docker build --rm -t ${imageTag} \
+sudo docker build ${REMOVE_CACHE_OPTION} -t ${imageTag} \
     ${BUILD_ARGS} \
     ${options} \
-    -f ${DOCKERFILE} .
+    -f $(basename ${DOCKERFILE}) .
+set +x
+cd -
+
+###################################################
+#### --------- More Guides for Users -------- #####
+###################################################
 
 echo "----> Shell into the Container in interactive mode: "
 echo "  docker exec -it --name <some-name> /bin/bash"
@@ -135,5 +200,5 @@ echo "----> Build Docker Images again: "
 echo "To build again: (there is a dot at the end of the command!)"
 echo "  docker build -t ${imageTag} . "
 echo
-docker images |grep "$imageTag"
+sudo docker images |grep "$imageTag"
